// lib/create-request-types.ts
export type RequestFamily =
  | "design_production"
  | "website_update"
  | "managed_communications";

export type RequestType =
  | "catalog_project"
  | "brochure_project"
  | "flyer_project"
  | "postcard_project"
  | "banner_project"
  | "fact_sheet_project"
  | "website_update"
  | "newsletter_request"
  | "social_media_request"
  | "campaign_support_request"
  | "other";